Relax the filetype detection pattern
The previous pattern doesn't have the flexibility for a prefix of path.
For instance `:e .github/workflows/ci.yml` is recognized as `yaml.gha`
filetype, but `:e ./.github/workflows/ci.yml` or
`:e /full/path/to/repo/.github/workflows/ci.yml` are not. However, these
commands edit the same file and it should be recognized as a
configuration file for GHA.