Fix mismatching to `}}` contained in a string
For instance, `${{ format('{{ hoge }}') }}` is recognized as a dollar
syntax(`${{ format('{{ hoge }}`) and a scalar(`') }}`) even though it
should be recognized as a single dollar syntax. This is due to the
`keepend` argument, it matches the first encountered end pattern even if
it is contained in another syntax item if the argument is given. So we
remove it to fix this problem.